+++ packages/core/src/transports/base.ts
@@ -3,7 +3,11 @@ import { BaseExtension } from '../extensions';
 import type { Transport, TransportItem } from './types';
 
 export abstract class BaseTransport extends BaseExtension implements Transport {
-  abstract send(items: TransportItem[]): void | Promise<void>;
+  abstract send(items: TransportItem | TransportItem[]): void | Promise<void>;
+
+  isBatched(): boolean {
+    return false;
+  }
 
   getIgnoreUrls(): Array<string | RegExp> {
     return [];